Do rudimentary route handling + controller work. Need auth + refactor for incoming db
